Don't fail on existing directories with SF_IMMUTABLE set
When updating a directory that has SF_IMMUTABLE set, unlike files,
libarchive is not good about clearing the flag while while it makes
changes. Work around this by manually clearing it ourselves if set;
libarchive will restore it if the archive entry for the new directory
requires it.

This went unnoticed due to FreeBSD's build system failing to set schg
for /var/empty when building disk images and distribution tarballs on
Linux.